# src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "2",
    "title" : "Hydroserving Manager service",
    "termsOfService" : "",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://github.com/Hydrospheredata/hydro-serving/blob/master/LICENSE"
    }
  },
  "basePath" : "/api/v2",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/application/{applicationName}" : {
      "delete" : {
        "tags" : [ "Application" ],
        "summary" : "deleteApplication",
        "description" : "deleteApplication",
        "operationId" : "deleteApplication",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "applicationName",
          "in" : "path",
          "description" : "name",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Application Deleted"
          }
        }
      }
    },
    "/application/generateInputs/{applicationName}" : {
      "get" : {
        "tags" : [ "Application" ],
        "summary" : "Generate payload for application",
        "description" : "Generate payload for application",
        "operationId" : "Generate payload for application",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "applicationName",
          "in" : "path",
          "description" : "name",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Any"
          }
        }
      }
    },
    "/application" : {
      "get" : {
        "tags" : [ "Application" ],
        "summary" : "applications",
        "description" : "applications",
        "operationId" : "applications",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Application",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ApplicationView"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Application" ],
        "summary" : "Add Application",
        "description" : "Add Application",
        "operationId" : "addApplication",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Application",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CreateApplicationRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Application",
            "schema" : {
              "$ref" : "#/definitions/ApplicationView"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Application" ],
        "summary" : "Update Application",
        "description" : "Update Application",
        "operationId" : "updateApplication",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "ApplicationCreateOrUpdateRequest",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UpdateApplicationRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Application",
            "schema" : {
              "$ref" : "#/definitions/ApplicationView"
            }
          }
        }
      }
    },
    "/application/{appName}" : {
      "get" : {
        "tags" : [ "Application" ],
        "summary" : "getApp",
        "description" : "getApp",
        "operationId" : "getApp",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "name",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Application",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ApplicationView"
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : { },
  "definitions" : {
...
    "DataProfileType" : {
      "type" : "object",
      "required" : [ "audio", "categorical", "continuous", "image", "interval", "nominal", "none", "numerical", "ordinal", "ratio", "text", "unrecognized", "video" ],
      "properties" : {
        "text" : {
          "type" : "boolean"
        },
        "none" : {
          "type" : "boolean"
        },
        "audio" : {
          "type" : "boolean"
        },
        "image" : {
          "type" : "boolean"
        },
        "video" : {
          "type" : "boolean"
        },
        "categorical" : {
          "type" : "boolean"
        },
        "nominal" : {
          "type" : "boolean"
        },
        "ordinal" : {
          "type" : "boolean"
        },
        "numerical" : {
          "type" : "boolean"
        },
        "continuous" : {
          "type" : "boolean"
        },
        "interval" : {
          "type" : "boolean"
        },
        "ratio" : {
          "type" : "boolean"
        },
        "unrecognized" : {
          "type" : "boolean"
        }
      }
    },
    "DataType" : {
      "type" : "object",
      "required" : [ "dtBfloat16", "dtBool", "dtComplex128", "dtComplex64", "dtDouble", "dtFloat", "dtHalf", "dtInt16", "dtInt32", "dtInt64", "dtInt8", "dtInvalid", "dtMap", "dtQint16", "dtQint32", "dtQint8", "dtQuint16", "dtQuint8", "dtResource", "dtString", "dtUint16", "dtUint32", "dtUint64", "dtUint8", "dtVariant", "unrecognized" ],
      "properties" : {
        "dtInvalid" : {
          "type" : "boolean"
        },
        "dtFloat" : {
          "type" : "boolean"
        },
        "dtDouble" : {
          "type" : "boolean"
        },
        "dtInt32" : {
          "type" : "boolean"
        },
        "dtUint8" : {
          "type" : "boolean"
        },
        "dtInt16" : {
          "type" : "boolean"
        },
        "dtString" : {
          "type" : "boolean"
        },
        "dtBool" : {
          "type" : "boolean"
        },
        "dtComplex64" : {
          "type" : "boolean"
        },
        "dtInt64" : {
          "type" : "boolean"
        },
        "dtQint8" : {
          "type" : "boolean"
        },
        "dtQuint8" : {
          "type" : "boolean"
        },
        "dtQint32" : {
          "type" : "boolean"
        },
        "dtBfloat16" : {
          "type" : "boolean"
        },
        "dtQint16" : {
          "type" : "boolean"
        },
        "dtQuint16" : {
          "type" : "boolean"
        },
        "dtUint16" : {
          "type" : "boolean"
        },
        "dtComplex128" : {
          "type" : "boolean"
        },
        "dtHalf" : {
          "type" : "boolean"
        },
        "dtResource" : {
          "type" : "boolean"
        },
        "dtVariant" : {
          "type" : "boolean"
        },
        "dtUint32" : {
          "type" : "boolean"
        },
        "dtUint64" : {
          "type" : "boolean"
        },
        "dtMap" : {
          "type" : "boolean"
        },
        "dtInt8" : {
          "type" : "boolean"
        },
        "unrecognized" : {
          "type" : "boolean"
        }
      }
    },
    "Dim" : {
      "type" : "object",
      "required" : [ "allFields", "name", "size" ],
      "properties" : {
        "size" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "DockerImage" : {
      "type" : "object",
      "required" : [ "name", "tag" ],
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "tag" : {
          "type" : "string"
        },
        "sha256" : {
          "type" : "string"
        }
      }
    },
    "Internal" : {
      "type" : "object",
      "required" : [ "created", "id", "image", "metadata", "model", "modelContract", "modelVersion", "runtime", "status" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "image" : {
          "$ref" : "#/definitions/DockerImage"
        },
        "created" : {
          "type" : "integer",
          "format" : "int64"
        },
        "finished" : {
          "type" : "integer",
          "format" : "int64"
        },
        "modelVersion" : {
          "type" : "integer",
          "format" : "int64"
        },
        "modelContract" : {
          "$ref" : "#/definitions/ModelContract"
        },
        "runtime" : {
          "$ref" : "#/definitions/DockerImage"
        },
        "model" : {
          "$ref" : "#/definitions/Model"
        },
        "status" : {
          "$ref" : "#/definitions/Value"
        },
        "installCommand" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        }
      }
    },
...
    "Model" : {
      "type" : "object",
      "required" : [ "id", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "ModelContract" : {
      "type" : "object",
      "required" : [ "allFields", "modelName", "predict" ],
      "properties" : {
        "modelName" : {
          "type" : "string"
        },
        "predict" : {
          "$ref" : "#/definitions/ModelSignature"
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "ModelField" : {
      "type" : "object",
      "required" : [ "allFields", "dtype", "name", "profile", "shape", "subfields", "typeOrSubfields" ],
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "shape" : {
          "$ref" : "#/definitions/TensorShapeProto"
        },
        "profile" : {
          "$ref" : "#/definitions/DataProfileType"
        },
        "typeOrSubfields" : {
          "$ref" : "#/definitions/TypeOrSubfields"
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "subfields" : {
          "$ref" : "#/definitions/Subfield"
        },
        "dtype" : {
          "$ref" : "#/definitions/DataType"
        }
      }
    },
    "ModelSignature" : {
      "type" : "object",
      "required" : [ "allFields", "inputs", "outputs", "signatureName" ],
      "properties" : {
        "signatureName" : {
          "type" : "string"
        },
        "inputs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModelField"
          }
        },
        "outputs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModelField"
          }
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
...
    "Subfield" : {
      "type" : "object",
      "required" : [ "allFields", "data" ],
      "properties" : {
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModelField"
          }
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "TensorShapeProto" : {
      "type" : "object",
      "required" : [ "allFields", "dim", "unknownRank" ],
      "properties" : {
        "dim" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Dim"
          }
        },
        "unknownRank" : {
          "type" : "boolean"
        },
        "allFields" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
...
    "TypeOrSubfields" : {
      "type" : "object",
      "required" : [ "defined", "dtype", "empty", "subfields" ],
      "properties" : {
        "subfields" : {
          "type" : "boolean"
        },
        "dtype" : {
          "type" : "boolean"
        },
        "empty" : {
          "type" : "boolean"
        },
        "defined" : {
          "type" : "boolean"
        }
      }
    },
    "Value" : {
      "type" : "object"
    },
...
    "ApplicationGraphView" : {
      "type" : "object",
      "required" : [ "stages" ],
      "properties" : {
        "stages" : {
          "$ref" : "#/definitions/NonEmptyListStageView"
        }
      }
    },
    "ApplicationKafkaStream" : {
      "type" : "object",
      "required" : [ "destinationTopic", "sourceTopic" ],
      "properties" : {
        "sourceTopic" : {
          "type" : "string"
        },
        "destinationTopic" : {
          "type" : "string"
        },
        "consumerId" : {
          "type" : "string"
        },
        "errorTopic" : {
          "type" : "string"
        }
      }
    },
    "ApplicationView" : {
      "type" : "object",
      "required" : [ "executionGraph", "id", "kafkaStreaming", "metadata", "name", "signature", "status" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "signature" : {
          "$ref" : "#/definitions/ModelSignature"
        },
        "executionGraph" : {
          "$ref" : "#/definitions/ApplicationGraphView"
        },
        "kafkaStreaming" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ApplicationKafkaStream"
          }
        },
        "message" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        }
      }
    },
...
    "NonEmptyListStageView" : {
      "type" : "object",
      "required" : [ "head", "tail" ],
      "properties" : {
        "head" : {
          "$ref" : "#/definitions/StageView"
        },
        "tail" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/StageView"
          }
        }
      }
    },
    "NonEmptyListVariantView" : {
      "type" : "object",
      "required" : [ "head", "tail" ],
      "properties" : {
        "head" : {
          "$ref" : "#/definitions/VariantView"
        },
        "tail" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/VariantView"
          }
        }
      }
    },
    "StageView" : {
      "type" : "object",
      "required" : [ "modelVariants", "signature" ],
      "properties" : {
        "modelVariants" : {
          "$ref" : "#/definitions/NonEmptyListVariantView"
        },
        "signature" : {
          "$ref" : "#/definitions/ModelSignature"
        }
      }
    },
    "VariantView" : {
      "type" : "object",
      "required" : [ "modelVersion", "weight" ],
      "properties" : {
        "modelVersion" : {
          "$ref" : "#/definitions/Internal"
        },
        "weight" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ExecutionGraphRequest" : {
      "type" : "object",
      "required" : [ "stages" ],
      "properties" : {
        "stages" : {
          "$ref" : "#/definitions/NonEmptyListPipelineStageRequest"
        }
      }
    },
    "ModelVariantRequest" : {
      "type" : "object",
      "required" : [ "modelVersionId", "weight" ],
      "properties" : {
        "modelVersionId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "weight" : {
          "type" : "integer",
          "format" : "int32"
        },
        "deploymentConfigName" : {
          "type" : "string"
        }
      }
    },
    "NonEmptyListModelVariantRequest" : {
      "type" : "object",
      "required" : [ "head", "tail" ],
      "properties" : {
        "head" : {
          "$ref" : "#/definitions/ModelVariantRequest"
        },
        "tail" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModelVariantRequest"
          }
        }
      }
    },
    "NonEmptyListPipelineStageRequest" : {
      "type" : "object",
      "required" : [ "head", "tail" ],
      "properties" : {
        "head" : {
          "$ref" : "#/definitions/PipelineStageRequest"
        },
        "tail" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PipelineStageRequest"
          }
        }
      }
    },
    "PipelineStageRequest" : {
      "type" : "object",
      "required" : [ "modelVariants" ],
      "properties" : {
        "modelVariants" : {
          "$ref" : "#/definitions/NonEmptyListModelVariantRequest"
        }
      }
    },
    "UpdateApplicationRequest" : {
      "type" : "object",
      "required" : [ "executionGraph", "id", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "namespace" : {
          "type" : "string"
        },
        "executionGraph" : {
          "$ref" : "#/definitions/ExecutionGraphRequest"
        },
        "kafkaStreaming" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ApplicationKafkaStream"
          }
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        }
      }
    },
    "CreateApplicationRequest" : {
      "type" : "object",
      "required" : [ "executionGraph", "name" ],
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "namespace" : {
          "type" : "string"
        },
        "executionGraph" : {
          "$ref" : "#/definitions/ExecutionGraphRequest"
        },
        "kafkaStreaming" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ApplicationKafkaStream"
          }
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        }
      }
    },
...
  },
...
}